backend/api/requirements.txt
# api
fastapi>=0.85.0
aiofiles
python-multipart
#db
sse-starlette
## utils
requests
python-magic
pydantic_settings
secure==0.3.0
tldextract
black==22.12.0
cffi==1.15.1
click==8.1.3
cryptography==41.0.0
flake8==6.0.0
h11==0.13.0
idna==3.4
isort==5.11.4
pre-commit==2.21.0
pycparser==2.21
python-dotenv==0.21.0
sniffio==1.3.0
setuptools>=65.5.1
jinja2
asgiref